Add ProtectedRoutes for Security
