+++ frontend/src/pages/Dashboard.jsx
@@ -3,19 +3,7 @@ import { useNavigate } from "react-router-dom";
 import DashboardNavigation from "../components/layout/DashboardNavigation";
 
 function DashboardPage() {
-  const navigate = useNavigate();
-  const [userData, setUserData] = useState(
-    JSON.parse(localStorage.getItem("userData"))
-  );
-
-  // Redirect to login if no user data is found
-  useEffect(() => {
-    if (!userData) {
-      navigate("/", { replace: true });
-    }
-  }, [userData, navigate]);
-
-  if (!userData) return null; // Prevents rendering if being redirected
+  const userData = JSON.parse(localStorage.getItem("userData"));
 
   return (
     <div>
